+++ cms-project/admin/includes/post_comments.php
@@ -81,7 +81,7 @@
       echo"<td>$comment_date</td>";
       echo"<td><a href='comments.php?source=post_comments&p_id={$comment_post_id}&approve={$comment_id}'>Approve</a></td>";
       echo"<td><a href='comments.php?source=post_comments&p_id={$comment_post_id}&deny={$comment_id}'>Deny</a></td>";
-      echo"<td><a href='comments.php?source=post_comments&p_id={$comment_post_id}&delete={$comment_id}'>Delete</a></td>";
+      echo"<td><a onClick=\"javascript: return confirm('Are you sure you want to delete this comment') \" href='comments.php?source=post_comments&p_id={$comment_post_id}&delete={$comment_id}'>Delete</a></td>";
     echo "</tr>";
     } // End while loop ?>
   </tbody>
